Extended the login page
The login page has an extra of 4 new sections, each section targetting a different aspect found in the area of interes of our site. One section has a container div, containing 2 divs, one that has an image, and one that has a description to that image, to the left or right of the image.